backend/requirements.txt
beautifulsoup4=="*"
requests=="*"
pandas=="*"
numpy=="*"
flask=="*"
flask-cors=="*"
werkzeug=="*"
flasgger=="*"
graphdb-python=="*"
duckduckgo_search=="*"